Update Dockerfile to reduce image size
I have implemented a multi-stage build, as per https://codefresh.io/docker-tutorial/create-docker-images-for-java/

The "FROM openjdk:8-jre-alpine3.9" on line 9 will discard the unneeded parts of the maven build. The alpine versions of maven and java are chosen because they are the smallest. This change reduced my image size from about 750 MB to only 120 MB. Build time was also reduced from about 100 seconds to only 42 seconds.

A possible downside to this change is that the latest versions of java and maven are not being used.